fix(regulated_report): correct use of annotations (#50)
We had a bug where by the compare pipeline wouldn't function without
provided EMBL annotations file.  This prevented use of automatically
generated windows being used.  We also fix the code path in the gene
statistics function so that we can use the appropriate logic depending
on if the EMBL file is one with actual padded annotations or one with
just the windows.